requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.63.0
fpdf==1.7.2
h11==0.12.0
img2pdf==0.4.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
pdf2image==1.14.0
pikepdf==2.11.3
Pillow==8.2.0
pydantic==1.8.1
PyPDF2==1.26.0
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
